+++ spec/api/endpoints/strava_endpoint_spec.rb
@@ -77,6 +77,55 @@ describe Api::Endpoints::StravaEndpoint do
             expect(response['ok']).to be true
           end
         end
+        context 'with multiple connected users with the same athlete id' do
+          let!(:team2) { Fabricate(:team) }
+          let!(:user2) do
+            Fabricate(
+              :user,
+              team: team2,
+              athlete: Fabricate.build(:athlete, athlete_id: user.athlete.athlete_id.to_s),
+              access_token: 'token'
+            )
+          end
+          it 'syncs both users' do
+            users = []
+            allow_any_instance_of(User).to receive(:sync_and_brag!) do |a_user, _args|
+              users << a_user
+            end
+            post '/api/strava/event',
+                 JSON.dump(
+                   event_data.merge(
+                     aspect_type: 'create',
+                     owner_id: user.athlete.athlete_id.to_s
+                   )
+                 ),
+                 'CONTENT_TYPE' => 'application/json'
+            expect(last_response.status).to eq 200
+            response = JSON.parse(last_response.body)
+            expect(response['ok']).to be true
+            expect(users).to eq([user, user2])
+          end
+          it 'skips over failures' do
+            users = []
+            allow_any_instance_of(User).to receive(:sync_and_brag!) do |a_user, _args|
+              raise 'error' if a_user == user
+
+              users << a_user
+            end
+            post '/api/strava/event',
+                 JSON.dump(
+                   event_data.merge(
+                     aspect_type: 'create',
+                     owner_id: user.athlete.athlete_id.to_s
+                   )
+                 ),
+                 'CONTENT_TYPE' => 'application/json'
+            expect(last_response.status).to eq 200
+            response = JSON.parse(last_response.body)
+            expect(response['ok']).to be true
+            expect(users).to eq([user2])
+          end
+        end
         context 'with an existing activity' do
           let!(:activity) { Fabricate(:user_activity, user: user, map: nil) }
           it 'rebrags the existing activity' do
@@ -131,6 +180,58 @@ describe Api::Endpoints::StravaEndpoint do
             response = JSON.parse(last_response.body)
             expect(response['ok']).to be true
           end
+          context 'with multiple connected users with the same athlete id' do
+            let!(:team2) { Fabricate(:team) }
+            let!(:user2) do
+              Fabricate(
+                :user,
+                team: team2,
+                athlete: Fabricate.build(:athlete, athlete_id: user.athlete.athlete_id.to_s),
+                access_token: 'token'
+              )
+            end
+            let!(:activity2) { Fabricate(:user_activity, user: user2, strava_id: activity.strava_id, map: nil) }
+            it 'rebrags both activities' do
+              users = []
+              allow_any_instance_of(User).to receive(:rebrag_activity!) do |a_user, _args|
+                users << a_user
+              end
+              post '/api/strava/event',
+                   JSON.dump(
+                     event_data.merge(
+                       aspect_type: 'update',
+                       object_id: activity.strava_id,
+                       owner_id: user.athlete.athlete_id.to_s
+                     )
+                   ),
+                   'CONTENT_TYPE' => 'application/json'
+              expect(last_response.status).to eq 200
+              response = JSON.parse(last_response.body)
+              expect(response['ok']).to be true
+              expect(users).to eq([user, user2])
+            end
+            it 'skips over failures' do
+              users = []
+              allow_any_instance_of(User).to receive(:rebrag_activity!) do |a_user, _args|
+                raise 'error' if a_user == user
+
+                users << a_user
+              end
+              post '/api/strava/event',
+                   JSON.dump(
+                     event_data.merge(
+                       aspect_type: 'update',
+                       object_id: activity.strava_id,
+                       owner_id: user.athlete.athlete_id.to_s
+                     )
+                   ),
+                   'CONTENT_TYPE' => 'application/json'
+              expect(last_response.status).to eq 200
+              response = JSON.parse(last_response.body)
+              expect(response['ok']).to be true
+              expect(users).to eq([user2])
+            end
+          end
           context 'a bragged activity' do
             before do
               activity.update_attributes!(channel_message: Fabricate(:channel_message))
@@ -150,6 +251,58 @@ describe Api::Endpoints::StravaEndpoint do
               response = JSON.parse(last_response.body)
               expect(response['ok']).to be true
               expect(activity.reload.channel_message).to be_nil
+            end
+            context 'with multiple connected users with the same athlete id' do
+              let!(:team2) { Fabricate(:team) }
+              let!(:user2) do
+                Fabricate(
+                  :user,
+                  team: team2,
+                  athlete: Fabricate.build(:athlete, athlete_id: user.athlete.athlete_id.to_s),
+                  access_token: 'token'
+                )
+              end
+              let!(:activity2) { Fabricate(:user_activity, user: user2, strava_id: activity.strava_id, map: nil) }
+              it 'deletes both activities' do
+                activities = []
+                allow_any_instance_of(UserActivity).to receive(:unbrag!) do |a_activity, _args|
+                  activities << a_activity
+                end
+                post '/api/strava/event',
+                     JSON.dump(
+                       event_data.merge(
+                         aspect_type: 'delete',
+                         object_id: activity.strava_id,
+                         owner_id: user.athlete.athlete_id.to_s
+                       )
+                     ),
+                     'CONTENT_TYPE' => 'application/json'
+                expect(last_response.status).to eq 200
+                response = JSON.parse(last_response.body)
+                expect(response['ok']).to be true
+                expect(activities).to eq([activity, activity2])
+              end
+              it 'skips over failures' do
+                activities = []
+                allow_any_instance_of(UserActivity).to receive(:unbrag!) do |a_activity, _args|
+                  raise 'error' if a_activity._id == activity._id
+
+                  activities << a_activity
+                end
+                post '/api/strava/event',
+                     JSON.dump(
+                       event_data.merge(
+                         aspect_type: 'delete',
+                         object_id: activity.strava_id,
+                         owner_id: user.athlete.athlete_id.to_s
+                       )
+                     ),
+                     'CONTENT_TYPE' => 'application/json'
+                expect(last_response.status).to eq 200
+                response = JSON.parse(last_response.body)
+                expect(response['ok']).to be true
+                expect(activities).to eq([activity2])
+              end
             end
           end
         end